==> enformer_padding_effect/1_enformer_torch_vs_tf.ipynb <==
import sys
sys.path.append("..")
from MPRA_predict.utils import *
from MPRA_predict.datasets import *

from torch.utils.data import DataLoader
from enformer_pytorch import Enformer, from_pretrained

import tensorflow as tf
import tensorflow_hub as hub
import gc

model = from_pretrained('../pretrained_models/enformer_weights', use_tf_gamma=True)

dataset = SeqLabelDataset(
    data_path='data/enformer_sequences_test_100.csv',
    data_type='seq',
    input_column='seq',
    crop=False,
    padding=False,
    N_fill_value=0)
test_data_loader = DataLoader(dataset, batch_size=4, shuffle=False, num_workers=4, pin_memory=True)
pred = get_pred(model, test_data_loader)

np.save(f'data/Enformer_pred_use_tf_gamma_true.npy', pred)

model = from_pretrained('../pretrained_models/enformer_weights', use_tf_gamma=False)

dataset = SeqLabelDataset(
    data_path='data/enformer_sequences_test_100.csv',
    data_type='seq',
    input_column='seq',
    crop=False,
    padding=False,
    N_fill_value=0)
test_data_loader = DataLoader(dataset, batch_size=4, shuffle=False, num_workers=4, pin_memory=True)
pred = get_pred(model, test_data_loader)

np.save(f'data/Enformer_pred_use_tf_gamma_false.npy', pred)

model_tf = hub.load("../pretrained_models/enformer_tfhub").model

df = pd.read_csv('data/enformer_sequences_test_100.csv')
seqs = np.array(df['seq'])
seqs = pad_seqs(seqs, 393216)
onehots = strs2onehots(seqs, N_fill_value=0)
onehots_tf =  tf.convert_to_tensor(onehots, dtype=tf.float32)

batch_size = 4
pred = []
for i in tqdm(range(0, len(onehots_tf), batch_size)):
    batch_inputs = onehots_tf[i:i+batch_size]
    batch_pred = model_tf.predict_on_batch(batch_inputs)
    pred.append(batch_pred['human'])

pred = tf.concat(pred, axis=0)
print(pred.shape)  # [100, 896, 5313]

np.save(f'data/Enformer_pred_tensorflow.npy', pred)

pred_list = []
pred = np.load(f'data/Enformer_pred_use_tf_gamma_true.npy')
pred_list.append(pred)
pred = np.load(f'data/Enformer_pred_use_tf_gamma_false.npy')
pred_list.append(pred)
pred = np.load(f'data/Enformer_pred_tensorflow.npy')
pred_list.append(pred)
target = np.load('data/enformer_targets_test_100.npy')
pred_list.append(target)
for i in range(len(pred_list)):
    print(pred_list[i].shape)
pred_list = np.stack(pred_list)
# print(pred_list.shape)

print(np.corrcoef(pred_list.reshape(pred_list.shape[0], -1)))

# 结论：由于tensorflow和pytorch的某个函数实现不同
# pytorch enformer只有在seq_len=196608bp且use_tf_gamma=True才能得到和tensorflow enformer一样的效果

# 如果seq_len != 196608,应该如何调整位置编码

for target_length in [2]:
    model = from_pretrained('../pretrained_models/enformer_weights', use_tf_gamma=True, target_length=2)

    dataset = SeqLabelDataset(
        data_path='data/enformer_sequences_test_100.csv',
        data_type='seq',
        input_column='seq',
        crop=True,
        cropped_length=256,
        padding=False,
        N_fill_value=0)
    test_data_loader = DataLoader(dataset, batch_size=4, shuffle=False, num_workers=4, pin_memory=True)
    pred = get_pred(model, test_data_loader)

    torch.cuda.empty_cache()
    np.save(f'data/Enformer_pred_crop_256_use_tf_gamma_true.npy', pred)

for target_length in [2]:
    model = from_pretrained('../pretrained_models/enformer_weights', use_tf_gamma=False, target_length=2)

    dataset = SeqLabelDataset(
        data_path='data/enformer_sequences_test_100.csv',
        data_type='seq',
        input_column='seq',
        crop=True,
        cropped_length=256,
        padding=False,
        N_fill_value=0)
    test_data_loader = DataLoader(dataset, batch_size=4, shuffle=False, num_workers=4, pin_memory=True)
    pred = get_pred(model, test_data_loader)

    torch.cuda.empty_cache()
    np.save(f'data/Enformer_pred_crop_256_use_tf_gamma_false.npy', pred)

pred_list = []
pred = np.load(f'data/Enformer_pred_crop_256_use_tf_gamma_true.npy')
pred_list.append(pred)
pred = np.load(f'data/Enformer_pred_crop_256_use_tf_gamma_false.npy')
pred_list.append(pred)
pred = np.load(f'data/Enformer_pred_use_tf_gamma_true.npy')[:, 447:449]
pred_list.append(pred)
pred = np.load(f'data/Enformer_pred_use_tf_gamma_false.npy')[:, 447:449]
pred_list.append(pred)
pred = np.load(f'data/Enformer_pred_tensorflow.npy')[:, 447:449]
pred_list.append(pred)
for i in range(len(pred_list)):
    print(pred_list[i].shape)
pred_list = np.stack(pred_list)
print(pred_list.shape)

print(np.corrcoef(pred_list.reshape(pred_list.shape[0], -1)))
