==> enformer_padding_effect/4_enformer_rc_aug.ipynb <==
# 实验5：augmentation (rc, shift) 的影响

import sys
sys.path.append("../../")

from torch.utils.data import DataLoader
from enformer_pytorch import from_pretrained

from MPRA_predict.utils import *
from MPRA_predict.datasets import *

np.set_printoptions(linewidth=120)

# 5.1 reverse complement

# df = pd.read_csv('data/enformer_sequences_test_100.csv')
# dataset = SeqLabelDataset(
#     data_df=df,
#     data_type='bed',
#     genome_path='../../../../genome/hg38.fa',
#     crop=True, 
#     crop_method='center',
#     cropped_length=65536,
#     padding=True, 
#     padding_method='N',
#     padded_length=131072,
#     N_fill_value=0)
# print(dataset[0]['seq'].shape)

# only rc, crop

df = pd.read_csv('data/enformer_sequences_test_100.csv')
df['seq'] = seqs_rc(df['seq'])

for cropped_length in [256, 1024, 4096, 16384, 65536, 196608]:
    model = from_pretrained('../../pretrained_models/enformer_weights', target_length=2)
    dataset = SeqLabelDataset(
        data_df=df,
        data_type='seq',
        input_column='seq', 
        crop=True, 
        crop_method='center',
        cropped_length=cropped_length,
        padding=False, 
        padding_method='N',
        padded_length=196608,
        N_fill_value=0)
    
    test_data_loader = DataLoader(dataset, batch_size=4, shuffle=False, num_workers=1)
    y_pred = get_pred(model, test_data_loader)

    torch.cuda.empty_cache()
    np.save(f'data/Enformer_pred_crop_{cropped_length}_only_rc.npy', y_pred)

pred_list = []
for cropped_length in [256, 1024, 4096, 16384, 65536, 196608]:
    pred = np.load(f'data/Enformer_pred_crop_{cropped_length}_only_rc.npy')
    # print(pred.shape)
    pred_list.append(pred)
pred_list = np.stack(pred_list)
pred_list = pred_list[:, :, ::-1, :] # reverse

target = np.load('data/enformer_targets_test_100.npy')
pred_list = np.concatenate([pred_list, target[None, :]], axis=0)
print(pred_list.shape)

corr = np.corrcoef(pred_list.reshape(pred_list.shape[0], -1))
print(corr)

pred_list_only_rc = pred_list

pred_list = []
for cropped_length in [256, 1024, 4096, 16384, 65536, 196608]:
    pred = np.load(f'data/Enformer_pred_crop_{cropped_length}.npy')
    print(pred.shape)
    pred_list.append(pred)
pred_list = np.stack(pred_list)

target = np.load('data/enformer_targets_test_100.npy')
pred_list = np.concatenate([pred_list, target[None, :]], axis=0)
print(pred_list.shape)

corr = np.corrcoef(pred_list.reshape(pred_list.shape[0], -1))
print(corr)

pred_average = (pred_list_only_rc + pred_list) / 2
corr = np.corrcoef(pred_average.reshape(pred_average.shape[0], -1))
print(corr)

# only rc, crop + pad

df = pd.read_csv('data/enformer_sequences_test_100.csv')
df['seq'] = seqs_rc(df['seq'])

for cropped_length in [256, 1024, 4096, 16384, 65536, 196608]:
    model = from_pretrained('../../pretrained_models/enformer_weights', target_length=2)
    dataset = SeqLabelDataset(
        data_df=df,
        data_type='seq',
        input_column='seq', 
        crop=True, 
        crop_method='center',
        cropped_length=cropped_length,
        padding=True, 
        padding_method='N',
        padded_length=196608,
        N_fill_value=0)
    
    test_data_loader = DataLoader(dataset, batch_size=4, shuffle=False, num_workers=1)
    y_pred = get_pred(model, test_data_loader)

    torch.cuda.empty_cache()
    np.save(f'data/Enformer_pred_crop_{cropped_length}_pad_{196608}_only_rc.npy', y_pred)

pred_list = []
for cropped_length in [256, 1024, 4096, 16384, 65536, 196608]:
    pred = np.load(f'data/Enformer_pred_crop_{cropped_length}_pad_{196608}_only_rc.npy')
    pred_list.append(pred)
pred_list = np.stack(pred_list)
# reverse
pred_list = pred_list[:, :, ::-1, :]

target = np.load('data/enformer_targets_test_100.npy')
pred_list = np.concatenate([pred_list, target[None, :]], axis=0)
print(pred_list.shape)

corr = np.corrcoef(pred_list.reshape(pred_list.shape[0], -1))
print(corr)

pred_list_only_rc = pred_list

pred_list = []
for cropped_length in [256, 1024, 4096, 16384, 65536, 196608]:
    pred = np.load(f'data/Enformer_pred_crop_{cropped_length}_pad_{196608}.npy')
    print(pred.shape)
    pred_list.append(pred)
pred_list = np.stack(pred_list)

target = np.load('data/enformer_targets_test_100.npy')
pred_list = np.concatenate([pred_list, target[None, :]], axis=0)
print(pred_list.shape)

corr = np.corrcoef(pred_list.reshape(pred_list.shape[0], -1))
print(corr)

pred_average = (pred_list_only_rc + pred_list) / 2
corr = np.corrcoef(pred_average.reshape(pred_average.shape[0], -1))
print(corr)

# 需要调查，N和no 平均似乎效果更好？？
